+++ concordium-sdk/src/main/java/com/concordium/sdk/ClientV2.java
@@ -21,6 +21,8 @@ import com.concordium.sdk.responses.peerlist.Peer;
 import com.concordium.sdk.responses.peerlist.PeerInfo;
 import com.concordium.sdk.responses.rewardstatus.RewardsOverview;
 import com.concordium.sdk.responses.cryptographicparameters.CryptographicParameters;
+import com.concordium.sdk.responses.peerlist.Peer;
+import com.concordium.sdk.responses.peerlist.PeerInfo;
 import com.concordium.sdk.responses.transactionstatus.TransactionStatus;
 import com.concordium.sdk.transactions.AccountAddress;
 import com.concordium.sdk.transactions.AccountTransaction;
@@ -39,6 +41,7 @@ import lombok.var;
 import java.io.File;
 import java.io.IOException;
 import java.net.InetAddress;
+import java.net.InetSocketAddress;
 import java.net.UnknownHostException;
 import java.util.Iterator;
 import java.util.Optional;
@@ -397,6 +400,8 @@ public final class ClientV2 {
     }
 
     /**
+     * Returns a {@link ImmutableList} of {@link PeerInfo} containing information about the nodes' peers
+     *
      * Returns a {@link ImmutableList} of {@link PeerInfo} containing information about the nodes' peers.
      *
      * @return {@link ImmutableList} of {@link PeerInfo}
@@ -405,6 +410,38 @@ public final class ClientV2 {
     public ImmutableList<PeerInfo> getPeersInfo() throws UnknownHostException {
         var grpcOutput = this.server().getPeersInfo(Empty.newBuilder().build());
         return PeerInfo.parseToList(grpcOutput);
+    }
+
+    /**
+     * Tries to connect to a peer with the submitted {@link InetSocketAddress}.
+     * If successful, adds the peer to the list of given addresses.
+     * Note. The peer might not be connected instantly, in that case the node will
+     * try to establish the connection in near future. In this case {@link io.grpc.Status#OK} is returned
+     *
+     * @param socketAddress {@link InetSocketAddress} of the peer to connect to
+     * @throws {@link io.grpc.StatusRuntimeException} if unsuccessful
+     */
+    public void peerConnect(InetSocketAddress socketAddress) {
+        val grpcIpSocket = IpSocketAddress.newBuilder()
+                .setIp(IpAddress.newBuilder().setValue(socketAddress.getAddress().getHostAddress()).build())
+                .setPort(Port.newBuilder().setValue(socketAddress.getPort()))
+                .build();
+        this.server().peerConnect(grpcIpSocket);
+    }
+
+    /**
+     * Disconnect from the peer with the submitted {@link InetSocketAddress} and remove them from
+     * the address list if they are on it.
+     *
+     * @param socketAddress {@link InetSocketAddress} of the peer to disconnect from
+     * @throws {@link io.grpc.StatusRuntimeException} if unsuccessful
+     */
+    public void peerDisconnect(InetSocketAddress socketAddress) {
+        val grpcIpSocket = IpSocketAddress.newBuilder()
+                .setIp(IpAddress.newBuilder().setValue(socketAddress.getAddress().getHostAddress()).build())
+                .setPort(Port.newBuilder().setValue(socketAddress.getPort()))
+                .build();
+        this.server().peerDisconnect(grpcIpSocket);
     }
 
     /***